refactor: use css modules
- install sass
