+++ package.json
@@ -8,7 +8,8 @@
     "next": "latest",
     "prop-types": "^15.7.2",
     "react": "latest",
-    "react-dom": "latest"
+    "react-dom": "latest",
+    "sass": "^1.29.0"
   },
   "devDependencies": {
     "babel-plugin-inline-react-svg": "^1.0.1",
