+++ src/components/hero-image/hero-image.module.scss
@@ -1,0 +1,37 @@
+.imageSide {
+  position: relative;
+  width: 40%;
+  height: 100vh;
+  text-align: center;
+  line-height: 3.5rem;
+  font-size: 2rem;
+  background: linear-gradient(to top, #09203f 0%, #537895 100%);
+  color: white;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+}
+
+.heroImage {
+  position: absolute;
+  width: 90%;
+  top: 50%;
+  right: -200px;
+  transform: translateY(-50%);
+  box-shadow: 0 16px 48px rgba(45, 45, 45, 0.31);
+  border-radius: 6px;
+  opacity: 0;
+  overflow: hidden;
+  & > div {
+    position: initial !important;
+  }
+
+  & img {
+    object-fit: cover;
+  }
+
+  @media (min-width: 993px) {
+    opacity: 1;
+  }
+}
