+++ src/components/info-area/info-area.module.scss
@@ -1,0 +1,62 @@
+.infoAreaWrapper {
+  position: relative;
+  left: -25px;
+
+  h1 {
+    font-size: 50px;
+    margin-bottom: 15px;
+  }
+
+  a {
+    margin-left: 10px;
+    margin-right: 10px;
+    color: #537895;
+  }
+
+  span {
+    color: #6f6f6f;
+  }
+  p {
+    line-height: 27px;
+    margin-bottom: 15px;
+  }
+
+  @media (min-width: 993px) {
+    width: 50%;
+    left: 100px;
+
+    p {
+      max-width: 80%;
+    }
+  }
+}
+
+.trackingInExpand {
+  animation: tracking-in-expand 1s cubic-bezier(0.215, 0.61, 0.355, 1) both;
+}
+
+@-webkit-keyframes tracking-in-expand {
+  0% {
+    letter-spacing: -0.5em;
+    opacity: 0;
+  }
+  40% {
+    opacity: 0.6;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+
+@keyframes tracking-in-expand {
+  0% {
+    letter-spacing: -0.5em;
+    opacity: 0;
+  }
+  40% {
+    opacity: 0.6;
+  }
+  100% {
+    opacity: 1;
+  }
+}
