+++ src/components/social-icons/social-icons.module.scss
@@ -1,0 +1,12 @@
+.socialIcons {
+  position: absolute;
+  bottom: 30px;
+  right: 30px;
+  a {
+    margin-right: 10px;
+    svg {
+      width: 35px;
+      height: 35px;
+    }
+  }
+}
