+++ src/css/layout.css
@@ -81,147 +81,6 @@ html {
   background: #f4f4f4;
 }
 
-.main-title {
-  font-size: 50px;
-  margin-bottom: 15px;
-}
-
-.button {
-  overflow: hidden;
-
-  padding: 12px;
-
-  cursor: pointer;
-  user-select: none;
-  transition: all 150ms linear;
-  text-align: center;
-  white-space: nowrap;
-  text-decoration: none !important;
-  text-transform: none;
-  text-transform: capitalize;
-
-  color: #fff;
-  border: 0 none;
-  border-radius: 4px;
-
-  font-size: 13px;
-  font-weight: 500;
-  line-height: 1.3;
-
-  appearance: none;
-
-  justify-content: center;
-  align-items: center;
-  flex: 0 0 160px;
-}
-
-@media (min-width: 993px) {
-  button {
-    box-shadow: 2px 5px 10px #e4e4e4;
-  }
-}
-button:hover {
-  transition: all 150ms linear;
-
-  opacity: 0.85;
-}
-button:active {
-  transition: all 150ms linear;
-  opacity: 0.75;
-}
-button:focus {
-  outline: 1px dotted #959595;
-  outline-offset: -4px;
-}
-
-.-dark {
-  background: #333;
-}
-
-.info-area {
-  position: relative;
-  left: -25px;
-}
-
-.info-area a {
-  margin-left: 10px;
-  margin-right: 10px;
-  color: #537895;
-}
-
-.info-area span {
-  color: #6f6f6f;
-}
-.info-area p {
-  line-height: 27px;
-  margin-bottom: 15px;
-}
-
-@media (min-width: 993px) {
-  .info-area {
-    width: 50%;
-    left: 100px;
-  }
-
-  .info-area p {
-    max-width: 80%;
-  }
-}
-
-.hero-image {
-  position: absolute;
-  width: 90%;
-  top: 50%;
-  right: -200px;
-  transform: translateY(-50%);
-  box-shadow: 0 16px 48px rgba(45, 45, 45, 0.31);
-  border-radius: 6px;
-  opacity: 0;
-  overflow: hidden;
-}
-
-.hero-image > div {
-  position: initial !important;
-}
-
-.hero-image img {
-  object-fit: cover;
-}
-
-@media (min-width: 993px) {
-  .hero-image {
-    opacity: 1;
-  }
-}
-
-.tracking-in-expand {
-  animation: tracking-in-expand 1s cubic-bezier(0.215, 0.61, 0.355, 1) both;
-}
-@-webkit-keyframes tracking-in-expand {
-  0% {
-    letter-spacing: -0.5em;
-    opacity: 0;
-  }
-  40% {
-    opacity: 0.6;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-@keyframes tracking-in-expand {
-  0% {
-    letter-spacing: -0.5em;
-    opacity: 0;
-  }
-  40% {
-    opacity: 0.6;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
 .pageWrapper {
   max-width: 100%;
   max-height: 100vh;
@@ -231,42 +90,13 @@ button:focus {
   justify-content: space-between;
 }
 
-.imageSide {
-  position: relative;
-  width: 40%;
-  height: 100vh;
-  text-align: center;
-  line-height: 3.5rem;
-  font-size: 2rem;
-  background: linear-gradient(to top, #09203f 0%, #537895 100%);
-  color: white;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
-
 .mailinfo {
   position: absolute;
   top: 30px;
   right: 30px;
 }
+
 .mailinfo a {
   color: #333;
   text-decoration: none;
-}
-
-.socialIcons {
-  position: absolute;
-  bottom: 30px;
-  right: 30px;
-}
-
-.socialIcons a {
-  margin-right: 10px;
-}
-
-.socialIcons a svg {
-  width: 35px;
-  height: 35px;
 }
